% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics.R
\name{ly_analytics_details}
\alias{ly_analytics_details}
\title{Get post analytics details}
\usage{
ly_analytics_details(token, url, start = NULL, end = NULL,
  verbose = FALSE)
}
\arguments{
\item{token}{your token as returned by \code{\link{ly_token}}.}

\item{url}{URL of post to fetch details on. This must be in canonical form,
including \code{http://} scheme. see examples}

\item{start}{start of date range to consider traffic from,
limited to most recent 90 days. Defaults to \code{Sys.Date()-7}.}

\item{end}{end of date range to consider traffic from.
Defaults to \code{Sys.Date()-7}.}

\item{verbose}{prints feedback in the console.}
}
\description{
Returns the post's metadata, as well as total views and visitors in the
metrics field. By default, this returns the total pageviews on the link
for the last 90 days.
}
\examples{
\dontrun{
token <- ly_token("agenda.weforum.org", "XXxxX00X0X000XxXxXx000X0X0X00X")

# get all posts
posts <- ly_analytics(token)

# get details of random post
post_details <- ly_analytics_details(token, url = sample(posts$url, 1))
}

}
