% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics.R
\name{ly_analytics_meta_details}
\alias{ly_analytics_meta_details}
\title{Get meta analytics details}
\usage{
ly_analytics_meta_details(token, meta, value, pub.start = Sys.Date() - 7,
  pub.end = Sys.Date(), period.start = NULL, period.end = NULL,
  verbose = FALSE)
}
\arguments{
\item{token}{your token as returned by \code{\link{ly_token}}.}

\item{meta}{type of \code{value}, see details.}

\item{value}{value of \code{meta} to fetch detaisl of.}

\item{pub.start, pub.end}{publication filter start and end date.}

\item{period.start}{start and end of period to consider traffic from.}

\item{verbose}{prints feedback in the console.}

\item{pub.start, pub.end}{publication filter start and end date.}
}
\description{
Get detailed analytics on a meta object.
}
\details{
valid meta:
\itemize{
  \item{\code{author}}
  \item{\code{section}}
  \item{\code{tag}}
}
}
\examples{
\dontrun{
token <- ly_token("agenda.weforum.org", "XXxxX00X0X000XxXxXx000X0X0X00X")

# get 100 authors
authors <- ly_analytics(ly_token, type = "authors", verbose = TRUE)

# get author details
details <- ly_analytics_meta_details(token, "author", value = sample(authors$author, 1))
}

}
