% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referrers.R
\name{ly_referrers}
\alias{ly_referrers}
\title{Get referrers}
\usage{
ly_referrers(token, referrer.type = "social", section = NULL, tag = NULL,
  domain = NULL, start = Sys.Date() - 7, end = Sys.Date(),
  pub.start = NULL, pub.end = NULL, n = 100, verbose = FALSE)
}
\arguments{
\item{token}{your token as returned by \code{\link{ly_token}}.}

\item{referrer.type}{one of \code{social}, \code{search}, \code{other},
\code{internal}. Defaults to \code{social}.}

\item{section}{section to restrict this query to.}

\item{tag}{tag to restrict this query to.}

\item{domain}{domain to restrict this query to.}

\item{start}{start of date range to consider traffic from,
limited to most recent 90 days. Defaults to \code{Sys.Date()-7}.}

\item{end}{end of date range to consider traffic from.
Defaults to \code{Sys.Date()-7}.}

\item{pub.start}{publication filter start date, \code{posts} only.}

\item{pub.end}{publication filter end date, \code{posts} only.}

\item{n}{number of results to return}

\item{verbose}{prints feedback in the console}
}
\description{
Return a list of the top referrers for a referrer type. This can be either an
overall number, or a number restricted to a specific domain.
}
\examples{
\dontrun{
token <- ly_token("agenda.weforum.org", "XXxxX00X0X000XxXxXx000X0X0X00X")

refs <- ly_referrers(token)
gender <- ly_referrers(token, referrer.type = "search", tag = "gender-parity")
}

}
