% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referrers.R
\name{ly_referrers_details_meta}
\alias{ly_referrers_details_meta}
\title{Get referrers meta details}
\usage{
ly_referrers_details_meta(token, referrer.type = "social",
  meta.type = "posts", meta.value, section = NULL, tag = NULL,
  domain = NULL, start = NULL, end = NULL, pub.start = NULL,
  pub.end = NULL, n = 100, verbose = FALSE)
}
\arguments{
\item{token}{your token as returned by \code{\link{ly_token}}.}

\item{referrer.type}{one of \code{social}, \code{search}, \code{other},
\code{internal}. Defaults to \code{social}.}

\item{meta.type}{One of \code{posts}, \code{author}, \code{section}, \code{tag}.}

\item{meta.value}{the value of the \code{meta.type}.}

\item{section}{section to restrict this query to.}

\item{tag}{tag to restrict this query to.}

\item{domain}{domain to restrict this query to.}

\item{start}{start of date range to consider traffic from,
limited to most recent 90 days. Defaults to \code{Sys.Date()-7}.}

\item{end}{end of date range to consider traffic from.
Defaults to \code{Sys.Date()-7}.}

\item{pub.start}{publication filter start date, \code{posts} only.}

\item{pub.end}{publication filter end date, \code{posts} only.}

\item{n}{number of results to return.}

\item{verbose}{prints feedback in the console.}
}
\description{
Returns a list of posts falling under the specified author or section.
}
\examples{
\dontrun{
token <- ly_token("my.domain.com", "XXxxX00X0X000XxXxXx000X0X0X00X")

refs <- ly_referrers(token)
rstats <- ly_referrers_details_meta(token)
}

}
