% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referrers.R
\name{ly_referrers_post}
\alias{ly_referrers_post}
\title{Get referrers post detail}
\usage{
ly_referrers_post(token, url, start = NULL, end = NULL,
  verbose = FALSE)
}
\arguments{
\item{token}{your token as returned by \code{\link{ly_token}}.}

\item{url}{url to fetch referrers of.}

\item{start}{start of date range to consider traffic from,
limited to most recent 90 days. Defaults to \code{Sys.Date()-7}.}

\item{end}{end of date range to consider traffic from.
Defaults to \code{Sys.Date()-7}.}

\item{verbose}{prints feedback in the console.}
}
\description{
Return a list of the top referrers for a given url.
}
\examples{
\dontrun{
token <- ly_token("my.domain.com", "XXxxX00X0X000XxXxXx000X0X0X00X")

refs <- ly_referrers_post(token, "https://www.r-project.org/")
}

}
