% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shares.R
\name{ly_shares}
\alias{ly_shares}
\title{Get shares}
\usage{
ly_shares(token, start = NULL, end = NULL, pub.start = NULL,
  pub.end = NULL, n = 100, verbose = FALSE)
}
\arguments{
\item{token}{your token as returned by \code{\link{ly_token}}.}

\item{start}{start of date range to consider traffic from,
limited to most recent 90 days.}

\item{end}{end of date range to consider traffic from.}

\item{pub.start}{publication filter start date, \code{posts} only.}

\item{pub.end}{publication filter end date, \code{posts} only.}

\item{n}{number of results to return}

\item{verbose}{prints feedback in the console}
}
\description{
Calls Parsely Analytics API
}
\examples{
\dontrun{
token <- ly_token("agenda.weforum.org", "XXxxX00X0X000XxXxXx000X0X0X00X")

shares <- ly_shares(token)
gender <- ly_shares(token, n = 1000, verbose = TRUE)
}

}
