% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shares.R
\name{ly_shares_details}
\alias{ly_shares_details}
\title{Get post shares detail}
\usage{
ly_shares_details(token, url, verbose = FALSE)
}
\arguments{
\item{token}{your token as returned by \code{\link{ly_token}}.}

\item{url}{canonical URL for which share counts should be retrieved.}

\item{verbose}{prints feedback in the console}
}
\description{
For a given canonical URL, return the total share counts across the top
social networks. The following table shows what each integer value returned
in the response contains.
}
\examples{
\dontrun{
token <- ly_token("my.domain.com", "XXxxX00X0X000XxXxXx000X0X0X00X")

# get shares
shares <- ly_shares(token)

# get details on random shares
details <- ly_shares_details(token, url = sample(shares$url, 1))
}

}
