% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shares.R
\name{ly_shares_details_meta}
\alias{ly_shares_details_meta}
\title{Get post shares meta detail}
\usage{
ly_shares_details_meta(token, meta = "author", value, pub.start = NULL,
  pub.end = NULL, n = 100, verbose = FALSE)
}
\arguments{
\item{token}{your token as returned by \code{\link{ly_token}}.}

\item{meta}{one of \code{author}, \code{section}, \code{tag}.}

\item{value}{value of \code{meta}.}

\item{pub.start}{publication filter start date, \code{posts} only.}

\item{pub.end}{publication filter end date, \code{posts} only.}

\item{n}{number of results to return}

\item{verbose}{prints feedback in the console.}
}
\description{
For a given canonical URL, return the total share counts across the top social networks. 
The following table shows what each integer value returned in the response contains.
}
\examples{
\dontrun{
token <- ly_token("my.domain.com", "XXxxX00X0X000XxXxXx000X0X0X00X")

# get details on random shares
details <- ly_shares_details_meta(token, meta = "author", value = "John Doe")
}

}
