% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\name{ly_token}
\alias{ly_token}
\title{Build token}
\usage{
ly_token(key, secret)
}
\arguments{
\item{key}{the API key specified in your Parse.ly JavaScript
tracker; Required for all requests.}

\item{secret}{this parameter is required for every request and should
contain the secret token obtainable from the
\href{https://dash.parsely.com/agenda.weforum.org/settings/api/}{API Settings}
 page.}
}
\description{
Build parse.ly token
}
