% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics.R
\name{ly_analytics}
\alias{ly_analytics}
\title{Get Anaqlytics}
\usage{
ly_analytics(token, type = "posts", start = Sys.Date() - 7,
  end = Sys.Date(), pub.start = NULL, pub.end = NULL, section = NULL,
  author = NULL, tag = NULL, sort = "views", n = 100, verbose = FALSE)
}
\arguments{
\item{token}{your token as returned by \code{\link{ly_token}}.}

\item{type}{one of \code{posts}, \code{authors}, \code{sections}, \code{tags},
\code{referrers}.}

\item{start}{start of date range to consider traffic from,
limited to most recent 90 days. Defaults to \code{Sys.Date()-7}.}

\item{end}{end of date range to consider traffic from.
Defaults to \code{Sys.Date()-7}.}

\item{pub.start}{publication filter start date, \code{posts} only.}

\item{pub.end}{publication filter end date, \code{posts} only.}

\item{section}{filter the current top posts by a specific section, \code{posts} only.}

\item{author}{filter the current top posts by a specific author, \code{posts} only.}

\item{tag}{filter the current top posts by a specific tag, \code{posts} only.}

\item{sort}{sort value; defaults to \code{views}. See
\href{https://www.parse.ly/help/api/available-metrics/}{available metrics}.}

\item{n}{number of results to return}

\item{verbose}{prints feedback in the console}
}
\description{
Calls Parsely Analytics API
}
