% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\name{ly_token}
\alias{ly_token}
\title{Build token}
\usage{
ly_token(key, secret)
}
\arguments{
\item{key}{the API key specified in your Parse.ly JavaScript
tracker; Required for all requests.}

\item{secret}{this parameter is required for every request and should
contain the secret token obtainable from the dashboard.}
}
\description{
Build parse.ly token
}
\examples{
\dontrun{
token <- ly_token("subdomain.domain.net", "XXxxX00X0X000XxXxXx000X0X0X00X")
}

}
